+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Force kill emulator if SIGKILL doesn't work
+
 ## [1.0.4] - 2023-11-09
 
 ### Fixed